First Upload : OpenOPC for Alstom S8000
OpenOPC library is useful to people who want to manage communication in Python.

There is an error in the Alstom S8000 code with the read and write function. So I have "adapted" like a pig the iwrite function in the library, file OpenOPC.py, near the lines 750.

Packaged with this upload, some examples of communication, especially RiyadhPWR.py

Enjoy.